requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
openai
ollama
numpy
